+++ src/main/java/nl/utwente/groove/grammar/aspect/AspectNode.java
@@ -134,7 +134,6 @@ public class AspectNode extends ANode implements AspectElement, Fixable {
      * and the label errors to the errors in this node.
      */
     public void addLabel(AspectLabel label) {
-        assert label.isFixed();
         assert getGraphRole() == label.getGraphRole();
         assert !isParsed();
         getNodeLabels().add(label);
@@ -416,7 +415,7 @@ public class AspectNode extends ANode implements AspectElement, Fixable {
      * this is not a product node.
      */
     public List<AspectNode> getArgNodes() {
-        testFixed(true);
+        assert isFixed();
         return this.argNodes;
     }